requirements.txt
Flask
tensorflow-cpu
Pillow
numpy
gunicorn